Manually added .libs and fully compile socket and client application both exe getting generated. Linux build is yet to be checked.

+++ MultiUserQuiz/QuestionTimer/QuestionTimer.cpp
@@ -1,6 +1,6 @@
 #include "QuestionTimer.h"
 
-QuestionTimer::QuestionTimer (int limit_ms, std::function<void ()> callback)
+QuestionTimer::QuestionTimer (long long limit_ms, std::function<void ()> callback)
     : time_limit_ms (limit_ms), timeout_callback (callback)
 { }
