+++ MultiUserQuiz/QuestionTimer/QuestionTimer.h
@@ -16,7 +16,7 @@ class QuestionTimer {
 
     std::chrono::steady_clock::time_point start_time;
     std::chrono::steady_clock::time_point end_time;
-    int time_limit_ms;                                      // holds the timelimit - in bullet mode it will be 15 sec, for timebound mode it will be total time of test.
+    long long time_limit_ms;                                // holds the timelimit - in bullet mode it will be 15 sec, for timebound mode it will be total time of test.
 
     std::atomic<bool> answered{false};                      // This will be set to true when the user answers the question and in Stop function - 
                                                             // but this will not be used in TimeBoundMode
@@ -29,7 +29,7 @@ class QuestionTimer {
 
  public:
  
-                QuestionTimer       (int limit_ms, std::function<void ()> callback);
+                QuestionTimer       (long long limit_ms, std::function<void ()> callback);
                 ~QuestionTimer      ();
 
      void       Start               ();
